Use lstat instead of access to check file existence
Some devices have a permanent /vendor symlink to /system/vendor in
ramdisk. As during multirom bootup neither system nor vendor partition
is mounted, so /vendor points to a non-existent target. Running access()
on this symlink results in error with NOENT (no such file or directory)
as it tries to dereference the target of symlink. Thus, the symlinking
of /mrom_enc/vendor to /vendor also fails since the link with same name
already exists.

Fix this by using lstat() which reports the existence of symlink (broken
or working) and real file properly.